+++ CHANGES.md
@@ -17,6 +17,10 @@ The released versions correspond to PyPI releases.
 * changed the default for `FakeFilesystem.shuffle_listdir_results` to `True` to reflect
   the real filesystem behavior
 
+### Fixes
+* fixed a deadlock in `shutil.copytree` if copying using an `shutil` function as
+  `copy_function` argument (see [#1235](../../issues/1235))
+
 ## [Version 5.10.0](https://pypi.python.org/pypi/pyfakefs/5.10.0) (2025-10-11)
 Adds official support for Python 3.14. Last minor version before the 6.0 release.